Move mount_options under `mod channel`
`channel` contains all the API for creating the FUSE device, mounting and
unmounting, so this belongs here.
